Manage dropping for `BIGNUM` and `BN_MONT_CTX` in Rust.
