+++ Cargo.toml
@@ -263,6 +263,7 @@ include = [
 name = "ring"
 
 [dependencies]
+libc = "0.2.20"
 untrusted = "0.3.2"
 
 [target.'cfg(any(target_os = "redox", unix))'.dependencies]
